Separate aaa auth question plugin classes (#1803)
* separate aaa authentication login classes to match to other plugins

* removed comments
